# Optimize cleanup of shards, used in resharding