--- a/lib/collection/src/collection/clean.rs
+++ b/lib/collection/src/collection/clean.rs
@@ -330,7 +330,7 @@ async fn clean_task(
             OperationWithClockTag::from(CollectionUpdateOperations::PointOperation(
                 crate::operations::point_ops::PointOperations::DeletePoints { ids },
             ));
-        if let Err(err) = shard.update_local(delete_operation, true).await {
+        if let Err(err) = shard.update_local(delete_operation, last_batch).await {
             return Err(CollectionError::service_error(format!(
                 "Failed to delete points from shard: {err}",
             )));