Optimize the background task that takes care of cleaning up shards, specifically deleting points from it that don't belong in it anymore.

This change makes the cleanup process a lot faster, which means it'll complete in less wall clock time.

The primary change is about how we select what points to delete. Before, we used a hash ring filter during the scroll request to only scroll points that don't belong into the shard anymore. Now, we simply scroll all point IDs and filter  the list after the scroll request. The key thing here is that the hash ring check is very expensive. Before, the hash ring check was done a lot more often across all segments in the shard. Now we only check each point ID once.

To summarize, this PR does two things:
1. move hash ring check outside of scroll operation
2. use wait=false on all delete operations, except the last one

I don't have a fancy graph, but I do have performance numbers. This is on a single node, with 5 million points, resharding from 1 to 2 shards, with 2 or 100 segments.

Time for shard cleanup to complete:

0. Before this PR:
    - 2 segments: 115 seconds
    - 100 segments: 72 seconds
1. With hash ring check outside scroll operation:
    - 2 segments: 35 seconds
    - 100 segments: 53 seconds
2. With wait=false:
    - 2 segments: 20 seconds
    - 100 segments: 43 seconds

### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

4. [x] Does your submission pass tests?
5. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
6. [x] Have you checked your code using `cargo clippy --all --all-features` command?