@agourlay 2025-02-28
Out of curiosity, why is a check using the Hashring so expensive?