@generall 2025-02-28
I am not entirely understand why number of comparisons is different. In scroll we do not check all points, but only points which are candidates to our limit+offset pool. Is that because we search top-N results in each segment independently?