@generall 2025-02-28
CPU usage before and after this fix: 
![image](https://github.com/user-attachments/assets/e23fec32-2e78-4a70-ab55-1670f0603a53)