@timvisee 2025-03-03
> Is that because we search top-N results in each segment independently?

Yes

> CPU usage before and after this fix:
> ![image](https://private-user-images.githubusercontent.com/1935623/418188377-e23fec32-2e78-4a70-ab55-1670f0603a53.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDA5OTM5MDYsIm5iZiI6MTc0MDk5MzYwNiwicGF0aCI6Ii8xOTM1NjIzLzQxODE4ODM3Ny1lMjNmZWMzMi0yZTc4LTRhNzAtYWI1NS0xNjcwZjA2MDNhNTMucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDMwMyUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAzMDNUMDkyMDA2WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ZWQyZTczOWNiYWE0N2EyNGVlYTFhOTg0Nzc0MTIxNGY4ZWZjOWFmOWU2NTU4YjVmMTcxNzM3OTgzMWJiM2YzZiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.K_lR4e_r56pa6tIukbRmnQQ84UEDDAGP7amgDZv9_f4)

Awesome! Thank you for sharing the results. The difference seems bigger than I expected.

I assume that part of the spikes in the new part of the graph are just because of indexing.